Deferring analysis of repo upon restoring previous session
When GitAhead restores the previous session, each repository gets fully started, including an analysis of the working copy. This induces a visible hang if the previous session has numerous tabs. This commit modifies the restoration procedure so that the analysis of each repository only starts the first time its tab is activated.

This may or may not be desirable. The fundamental problem is that this analysis is not threaded and freezes the UI.